Bugfix for extra space
If there is a trailing space in the neutrino command line arguments, neutrino will throw an error and exit to browser.
If valid vmc info is present this does not matter because the vmc argument is added after the space.
The issue is when a list without vmc info is used and NEUTRINO_LaunchOptions still has the '-vmc ' flag.
This just removes the extra space at the end to ensure that no trailing space is present.